+++ CHANGELOG.md
@@ -12,6 +12,11 @@ FIXED:
 
 ## 3.1.1 - 2019-04-05
 
+DEPRECATED:
+
+- `ToOneRelationship::omitWhenNotIncluded()`: Use `ToOneRelationship::omitDataWhenNotIncluded()`
+- `ToManyRelationship::omitWhenNotIncluded()`: Use `ToManyRelationship::omitDataWhenNotIncluded()`
+
 FIXED:
 
 - [#82](https://github.com/woohoolabs/yin/pull/82): Fix output for ToOneRelationship when the relationship doesn't contain any keys